VB-1207 Update database calls to use the latest reserved bookings (#169)
* VB-1207 Update database calls to use the latest reserved bookings

* VB-1207 Update database calls to use the latest reserved bookings, updates after
Dhiraj Shetty review

* VB-1207 Update database calls to use the latest reserved bookings, updates after
Dhiraj Shetty review

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/model/entity/projections/VisitRestrictionStats.kt
@@ -4,5 +4,5 @@ import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitRestriction
 
 interface VisitRestrictionStats {
   val visitRestriction: VisitRestriction
-  val count: Long
+  val count: Int
 }
